feat: Loosen transmission hierarchy restrictions (#2099)
Pending final decision on how transmission hierarchy should be
restricted, we need to increase max width to unblock Skatt. At the same
time, the max depth has been reduced to 20. As the transmission feature
is not widely used, future restrictions is deemed to be manageable to
roll out without too much breakage.