fix: Add ActorNames to LargeDataSetGenerator (#2113)
<!--- Provide a general summary of your changes in the Title above -->

## Description

LargeDataSetGenerator is outdated, it is missing the re-write of
ActorName
Needs to be fixed so we can re-populate YT01 after the migration for the
re-write got "stuck"/took forever


This adds seeding of actor names, it takes the parties list and selects
party based on dialogId so it matches how dialog selects party.
This ensures the same party is mapped to the same dialog. 
We then store this id/name in a ConcurrentDictionary so we can check for
the next time the same party comes up.
(Id/name combos can only be inserted once)

Between each loop we fetch all actor names from the previous round of
seeding
(This takes under 1 second for ~800k parties, and we just keep it in
memory)

## Related Issue(s)

- #2112

## Verification

- [ ] **Your** code builds clean without any errors or warnings
- [ ] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)