fix: Add user type per endpoint policy validation (#2540)
## Description

This adds a per-endpoint-policy validation to the user type validation
middleware, ensuring that each endpoint cannot be accessed unless the
correct user type for that endpoint is authenticated.

Changes include an update to the e2e tests, which conflated the claims
(in particular scopes) for service owner and end user tokens. Now this
is properly separated to closer mimic the actual maskinporten/id-porten
tokens.

## Related Issue(s)

- #2538 

## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [x] Relevant automated test added (if you find this hard, leave it and
we'll help out)