feat(auth): Add app instance delegation support (#2561)
## Description

This adds support for Altinn 3 app instance delegations. 

It expands and refactors `CollapseSubjectResources` - now renamed to the
more descriptive `ResolveDialogSearchAuthorization` - to also build a
list of essentially serviceowner label values that the adapter uses when
syncing Altinn instances. In a separate step, a query is performed to
find the list of dialog ids that contains these labels, which is passed
on to the list authorization prefilter as before.

This also includes a covering index of serviceowner-value, including the
dialogserviceownercontextid (which is identical to the dialogid). This
should allow for fast lookups of the dialog ids.

## Related Issue(s)

- #2544 

## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [x] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [x] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)

---------

Co-authored-by: Ole Jørgen Skogstad <skogstad@softis.net>