ci: rewrite deployment, verification and run of reindex job (#2880)
- Separate verification of reindex job running. Ensure that the
verification retries until it finds the execution. Might be a delay
between starting the job and fetching the execution. 2 min should be
moore than enough..
- Separate the deployment of the job and the execution of the job. Added
the job as a part of the deployment of apps so we can do dry-run and
verify deployment in each environment. No execution is run before
running the dispatch. This will ensure that the execution starts based
on the latest job deployed in that environment.
- Use the azure CLI action instead of bash so we don't need to add
extension etc.
- Update the documentation to match the rewrite