chore(perf): Add party column, GIN index to DialogSearch (#3046)
## Description

This adds a Party column to DialogSearch, a corresponding composite
index, and updates the search and indexing mechanisms.

WARNING! This contains a migration
`20251125122030_AddDialogSearchPartyColumn` that _must_ be run manually
in production/YT01 environments before deployment (with
__EFMigrationHistory manually updated). Should work fine in low-volume
environments as ATxx, TT02.

The second migration (updating views, stored functions) can be run as
part of the EF migration.

## Related Issue(s)

- n/a

## Verification

- [x] **Your** code builds clean without any errors or warnings
- [x] Manual testing done (required)
- [ ] Relevant automated test added (if you find this hard, leave it and
we'll help out)

## Documentation

- [ ] Documentation is updated (either in `docs`-directory, Altinnpedia
or a separate linked PR in
[altinn-studio-docs.](https://github.com/Altinn/altinn-studio-docs), if
applicable)

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>